If we are going to have separate components for the dialog and the dropzone, we might as well be able to call them separately for people who want a dropzone without a dialog
