+++ example/src/DropzoneAreaExample.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import {DropzoneArea} from 'material-ui-dropzone'
+
+export default function(props){
+  return (
+    <DropzoneArea />
+  )
+}
