--- examples/app/index.html
@@ -1,22 +1,0 @@
-<!doctype html>
-<html class="no-js" lang="">
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title>Material-UI Dropzone Example</title>
-    <meta name="description" content="">
-
-    <meta
-            name="viewport"
-            content="width=device-width, initial-scale=1, user-scalable=0, maximum-scale=1, minimum-scale=1"
-    >
-
-</head>
-
-<body>
-<div id="root"></div>
-</body>
-
-<script src="bundle.js"></script>
-
-</html>
